Moves cleanClone to lib/utils/
