+++ index.js
@@ -1,7 +1,7 @@
 const postcss = require('postcss');
 const Tidy = require('./Tidy');
 const { getGlobalOptions } = require('./lib/parse-options');
-const cleanClone = require('./lib/cleanClone');
+const cleanClone = require('./lib/utils/cleanClone');
 const tidyShorthandProperty = require('./lib/tidy-shorthand-property');
 const tidyProperty = require('./lib/tidy-property');
 const tidyFunction = require('./lib/tidy-function');
