+++ lib/tidy-function.js
@@ -1,5 +1,5 @@
 /* eslint-disable no-param-reassign */
-const cleanClone = require('./cleanClone');
+const cleanClone = require('./utils/cleanClone');
 
 /**
  * Replace `tidy-[span|offset]()` and `tidy-[span|offset]-full()` functions.
