+++ lib/tidy-shorthand-property.js
@@ -1,4 +1,4 @@
-const cleanClone = require('./cleanClone');
+const cleanClone = require('./utils/cleanClone');
 
 /**
  * Clean and trim shorthand property values.
