+++ lib/utils/cleanClone.js
